Cleans up, prepare Reddit Plugin
