+++ package.json
@@ -11,9 +11,6 @@
     "format": "npx prettier --write '**/*.ts'",
     "prepare": "npx husky install",
     "slash": "node build/utils/deploy-commands.js",
-    "k8s:clean": "kubectl delete -f ./k8s/namespace.yaml",
-    "k8s:deploy": "kubectl apply -f ./k8s/namespace.yaml && kubectl apply -f ./k8s/service.yaml && kubectl apply -f ./k8s/secrets.yaml && kubectl apply -f ./k8s/deployment.yaml",
-    "k8s:prod": "kubectl apply -f ./k8s/namespace.yaml && kubectl apply -f ./k8s/service.yaml && kubectl apply -f ./k8s/secrets.prod.yaml && kubectl apply -f ./k8s/deployment.yaml",
     "slashDev": "npm run build && node build/utils/deploy-commands.js",
     "test": "ts-mocha -p ./tsconfig.json ./tests/**/*.spec.ts"
   },
